src/chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Phil Andreoli <phil.andreoli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "1.31.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
